Size of RAM block chunks changed form 8 to 4 bytes

+++ project/src/headers/types.h
@@ -8,6 +8,6 @@ typedef unsigned long int 	addr_typ;
 
 
 // The type of the data in the memory
-typedef double data_typ;
+typedef float data_typ;
 
 #endif
